Screens and Debugging
Figured out how to tell myself what screen I'm on, and what to press to do what. Struggling on not getting stuck on minutiae. Also discovered that variable naming can be so annoying- my app.moneygoals gave me a headache because i was inconsistent with calling it app.moneygoal vs app.moneygoals.