Fixed Fries In Bag!
Had a weird width bug that was from me trying to use enumerate, I fixed it by just using a for loop. I then realized I had to use a while loop to not have my fries overlapping become a weird problem... not fully sure why the fix worked, but so far I haven't been able to break it!